# api/src/generated/src/main/java/com/workflow/api/PhoneContactsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.workflow.api;

import com.workflow.dto.PhoneContactDTO;
import com.workflow.dto.PhoneContactPageDTO;
import com.workflow.dto.PhoneContactUpdateRequestDTO;
import com.workflow.dto.PhoneUserDTO;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-07T10:14:28.616693700+01:00[Europe/Paris]", comments = "Generator version: 7.7.0")
@Validated
public interface PhoneContactsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /phone-contacts/{id} : Get le contact lié a l&#39;id
     *
     * @param id  (required)
     * @return Contact (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/phone-contacts/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PhoneUserDTO> getPhoneContact(
         @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"firstname\" : \"Jean\", \"phonePerso\" : { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"phoneOther\" : { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"phoneTelavox\" : { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"company\" : \"Saamp SA\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"phonePro\" : { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, \"lastname\" : \"Dupont\", \"denomination\" : \"ACME SAS\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /phone-contacts/page : Liste paginée des contacts téléphoniques (avec filtre)
     *
     * @param page  (required)
     * @param size  (required)
     * @param q Filtre texte (numéro, prénom, nom, société) (optional)
     * @return Page de contacts (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/phone-contacts/page",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PhoneContactPageDTO> getPhoneContactsPage(
        @NotNull @Min(0)  @Valid @RequestParam(value = "page", required = true, defaultValue = "0") Integer page,
        @NotNull @Min(1) @Max(200)  @Valid @RequestParam(value = "size", required = true, defaultValue = "25") Integer size,
         @Valid @RequestParam(value = "q", required = false) String q
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"meta\" : { \"size\" : 50, \"hasPrev\" : false, \"totalPages\" : 247, \"hasNext\" : true, \"page\" : 0, \"totalElements\" : 12345 }, \"items\" : [ { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, { \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /phone-contacts/{id} : Mettre à jour les informations d&#39;un contact téléphonique
     *
     * @param id  (required)
     * @param phoneContactUpdateRequestDTO  (required)
     * @return Contact mis à jour (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/phone-contacts/{id}",
        produces = { "application/json" },
        consumes = "application/json"
    )
    
    default ResponseEntity<PhoneContactDTO> updatePhoneContact(
         @PathVariable("id") UUID id,
         @Valid @RequestBody PhoneContactUpdateRequestDTO phoneContactUpdateRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"phoneNumber\" : \"33651440897\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}