# api/src/generated/src/main/java/com/workflow/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.workflow.api;

import java.util.UUID;
import com.workflow.dto.UserDTO;
import com.workflow.dto.UserSyncResponseDTO;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-07T10:14:28.616693700+01:00[Europe/Paris]", comments = "Generator version: 7.7.0")
@Validated
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/users/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/users/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserDTO> getUserById(
         @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"password\" : \"password\", \"erpId\" : 0, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"first\" : \"first\", \"email\" : \"email\", \"lastname\" : \"lastname\", \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/users/sync : Lancer la synchronisation des utilisateurs (Oracle -&gt; Postgres)
     * Importe les utilisateurs ayant un PC_COMP_ID valide. Met à jour les existants.
     *
     * @return Synchronisation terminée avec succès (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/users/sync",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserSyncResponseDTO> syncUsersFromLegacy(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"usersProcessed\" : 0, \"success\" : true, \"message\" : \"message\", \"groupsIgnored\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}