# api/src/main/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Saamp Web API
  version: 1.0.0
...
paths:
...
  /api/users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: './schemas/user.yaml#/User' } } } } # *
        '404': { description: Not found }

  # --- Endpoints Téléphonie ---
...
  /phone-contacts/page:
    get:
      tags: [PhoneContacts]
      summary: Liste paginée des contacts téléphoniques (avec filtre)
      operationId: getPhoneContactsPage
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
          description: Filtre texte (numéro, prénom, nom, société)
        - in: query
          name: page
          required: true
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          required: true
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
      responses:
        '200':
          description: Page de contacts
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneContactPage' }
  /phone-contacts/{id}:
    get:
      tags: [ PhoneContacts ]
      summary: Get le contact lié a l'id
      operationId: getPhoneContact
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Contact
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneUser' }
    put:
      tags: [PhoneContacts]
      summary: Mettre à jour les informations d'un contact téléphonique
      operationId: updatePhoneContact
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/telephony.yaml#/PhoneContactUpdateRequest' }
      responses:
        '200':
          description: Contact mis à jour
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneContact' }
  # --- Endpoints Tarifs (Pricing) ---
...